# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audio Downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoInfo {
  title: string;
  /** Duration in seconds */
  duration: number;
  /** Thumbnail URL */
  thumbnail: string;
  /**
     * Approximate filesize in bytes
     * @nullable
     */
  filesize?: number | null;
  /**
     * Audio bitrate in kbps
     * @nullable
     */
  bitrate?: number | null;
  videoId: string;
}

export type DownloadInputFormat = typeof DownloadInputFormat[keyof typeof DownloadInputFormat];


export const DownloadInputFormat = {
  mp3: 'mp3',
  m4a: 'm4a',
} as const;

export interface DownloadInput {
  url: string;
  format: DownloadInputFormat;
}

export interface DownloadJob {
  jobId: string;
}

export interface ApiError {
  error: string;
}

export type GetVideoInfoParams = {
/**
 * YouTube video URL
 */
url: string;
};