# lib/api-zod/src/generated/types/downloadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audio Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadInputFormat } from './downloadInputFormat';

export interface DownloadInput {
  url: string;
  format: DownloadInputFormat;
}