# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audio Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoInfo {
  title: string;
  /** Duration in seconds */
  duration: number;
  /** Thumbnail URL */
  thumbnail: string;
  /**
     * Approximate filesize in bytes
     * @nullable
     */
  filesize?: number | null;
  /**
     * Audio bitrate in kbps
     * @nullable
     */
  bitrate?: number | null;
  videoId: string;
}